opencl: Adreno 850 (art.api37 / E17 compiler) compat
The Adreno 850 ships a new shader compiler series, "E17", which the driver-string
parser does not recognise -- it classifies only E031 and DX. Everything keyed on
the parsed version therefore misbehaves on this part, and three of its kernels do
not build at all.

- Parse the E17 series. Its prefix is three characters, so each version field sits
  one byte earlier than in the E031 layout.
- mul_mv_q4_k_f32: add the missing "#pragma OPENCL EXTENSION cl_khr_fp16 : enable".
  E17 is spec-strict and rejects a direct load from a __global half* without it
  (err -11). Older Adreno compilers accepted it implicitly. mul_mv_q5_k_f32 already
  has the pragma and was never affected.
- gemm_moe_mxfp4_f32_ns: cast the e8m0 scale to half. A bare half8 * float is a
  vector-by-higher-rank-scalar multiply, which E17 rejects. OpenCL converts the
  scalar to the vector element type anyway, so the result is unchanged elsewhere.
- Build mul_mv_q4_0_f32_1d_8x_flat on E17. It was skipped as if it were an old E031
  compiler, yet the Adreno q4_0 GEMM dispatch (ne11 > 2) still selects it, so the
  first such matmul died with CL_INVALID_KERNEL.
- Decline FLASH_ATTN_EXT on E17: its compiler segfaults while building the flash_attn
  programs, taking the process with it. Attention falls back to the unfused path.

The FA decline is a correctness workaround for a compiler defect, so it is keyed on
the compiler class and stays on until a fixed compiler is verified, rather than
lifting itself on a version bump. GGML_OPENCL_ART_QUIRKS=0 disables the workarounds
in a single run so a new compiler can be re-verified without a rebuild.

Dense q4_0 and q4_K models run coherently on the Adreno 850 GPU. No change on any
other device: no other compiler parses as E17.

+++ ggml/src/ggml-opencl/kernels/gemm_moe_mxfp4_f32_ns.cl
@@ -274,8 +274,11 @@ kernel void kernel_gemm_moe_mxfp4_f32_ns(
         shared_b[b_local_offset.y] = bx8_f16.hi;
 
         // Dequantization
-        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * s;
-        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * s;
+        // Cast the e8m0 scale to half: a bare half8 * float is a vector-by-higher-rank-scalar
+        // multiply, which the art.api37 (E17) shader compiler rejects. OpenCL converts the
+        // scalar to the vector element type anyway, so this is identical on every device.
+        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * (half)s;
+        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * (half)s;
 
         sub_group_barrier(CLK_LOCAL_MEM_FENCE);
 
@@ -304,8 +307,11 @@ kernel void kernel_gemm_moe_mxfp4_f32_ns(
         shared_b[b_local_offset.y] = bx8_f16.hi;
 
         // Dequantization
-        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * s;
-        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * s;
+        // Cast the e8m0 scale to half: a bare half8 * float is a vector-by-higher-rank-scalar
+        // multiply, which the art.api37 (E17) shader compiler rejects. OpenCL converts the
+        // scalar to the vector element type anyway, so this is identical on every device.
+        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * (half)s;
+        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * (half)s;
 
         sub_group_barrier(CLK_LOCAL_MEM_FENCE);
